+++ CHANGES.md
@@ -1,5 +1,20 @@
 # Changes
 
+## 0.0.15
+
+- makes entry endpoint work by fixing #63
+
+## 0.0.14
+
+- fixes bugs #60 ... #62
+
+## 0.0.13
+
+- fixes issue issue with URL resource provider (see issue #57)
+- uses DTS Transformations 0.6.1 and thus not nesting dublinCore
+  metadata object by default any more (see issue
+  [scdh/dts-transformations#26](https://github.com/SCDH/dts-transformations/issues/26))
+
 ## 0.0.12
 
 - bundled with DTS Transformations 0.6.0